Create dedicated files for pages in App file
