+++ src/App.tsx
@@ -1,6 +1,9 @@
 import './App.css';
+import { useEffect } from 'react';
 import { Route, Routes } from 'react-router-dom';
+import { removeUserSession } from './Utils/Common';
 import Header from './components/Header'
+import Footer from './components/Footer'
 import Home from './pages/Home'
 import Products from './pages/Products'
 import Product from './pages/Product'
@@ -8,10 +11,10 @@ import NewProduct from './pages/NewProduct'
 import SignUp from './pages/SignUp'
 import SignIn from './pages/SignIn'
 import EditProduct from './pages/EditProduct';
-import { removeUserSession } from './Utils/Common';
 import Users from './pages/Users';
 import EditUser from './pages/EditUser';
-import { useEffect } from 'react';
+import { BadRequest, Unauthorized, AccessDenied, PageNotFound } from './pages/ErrorPages';
+import Logout from './pages/Logout';
 
 const App = () => (
   <div className="app">
@@ -20,58 +23,6 @@ const App = () => (
     <Footer />
   </div>
 );
-
-const Footer = () => (
-  <div className="footer">
-    <div className="text-info">Copyright © 2021 TripleS - Smart Storage System</div>
-  </div>
-);
-
-const PageNotFound = () => (
-  <div className="page-not-found">
-    <div className="text-center py-5">
-      <h1 className="display-1">404</h1>
-      <h2>Page not found</h2>
-    </div>
-  </div>
-);
-
-const AccessDenied = () => (
-  <div className="page-not-found">
-    <div className="text-center py-5">
-      <h1 className="display-1">403</h1>
-      <h2>Access Denied</h2>
-    </div>
-  </div>
-);
-
-const BadRequest = () => (
-  <div className="page-not-found">
-    <div className="text-center py-5">
-      <h1 className="display-1">400</h1>
-      <h2>Bad Request</h2>
-    </div>
-  </div>
-);
-
-function Logout() {
-  removeUserSession()
- 
-  useEffect(() => {
-    setTimeout(function() {
-      window.location.href="/"
-    }, 500)
-  }, [])
-
-  return (
-    <div className="page-not-found">
-      <div className="text-center py-5">
-        <h1 className="display-1">Bye</h1>
-        <h2>Log out</h2>
-      </div>
-    </div>
-  );
-}
 
 const Main = () => (
   <div className="main">
@@ -87,6 +38,7 @@ const Main = () => (
       <Route path='/signin' element={<SignIn />} />
       <Route path='/logout' element={<Logout />} />
       <Route path='/403' element={<AccessDenied />} />
+      <Route path='/401' element={<Unauthorized />} />
       <Route path='/400' element={<BadRequest />} />
       <Route path='*' element={<PageNotFound />} />
     </Routes>
