+++ src/components/Footer.tsx
@@ -1,0 +1,10 @@
+import '../App.css';
+
+
+const Footer = () => (
+  <div className="footer">
+    <div className="text-info">Copyright © 2021 TripleS - Smart Storage System</div>
+  </div>
+);
+
+export default Footer;
